# backend/api/rest/openapi.yaml
openapi: 3.0.3
info:
  title: OneBook AI Gateway API
  version: 0.1.0
  description: |
    Gateway REST API contract. All public requests go through the gateway.
    If the client sends `X-Request-Id`, the gateway propagates it end-to-end.
    If missing, the gateway generates one and returns it in both response header
    `X-Request-Id` and error payload field `requestId`.
servers:
  - url: http://localhost:8080
    description: Local dev (gateway)
...
paths:
...
  /api/books:
    get:
      tags: [books]
      summary: List books
      security:
        - sessionCookieAuth: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, processing, ready, failed]
        - name: primaryCategory
          in: query
          schema:
            type: string
            enum: [course_material, research_paper, project_doc, policy_regulation, reference_book, personal_note, how_to_guide, other]
        - name: tag
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [pdf, epub, txt]
        - name: language
          in: query
          schema:
            type: string
            enum: [zh, en, other, unknown]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListBooksResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [books]
      summary: Upload book
      description: |
        Upload a PDF/EPUB/TXT file. Size and allowed extensions are configurable
        on the server (default 50MB, .pdf/.epub/.txt).
      security:
        - sessionCookieAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                primaryCategory:
                  type: string
                  enum: [course_material, research_paper, project_doc, policy_regulation, reference_book, personal_note, how_to_guide, other]
                tags[]:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Replay hit
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                enum: ["true"]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: onebook_access
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Stable machine-readable error code (domain-specific, e.g. AUTH_INVALID_CREDENTIALS / BOOK_NOT_FOUND / CHAT_BOOK_NOT_READY).
        requestId:
          type: string
          description: Request correlation id, same as response header X-Request-Id.
        details:
          type: array
          description: Optional field-level validation details.
          items:
            $ref: "#/components/schemas/ErrorDetail"
      required: [error, code]
      example:
        error: book not ready
        code: CHAT_BOOK_NOT_READY
        requestId: req-demo-20260213-001
        details:
          - reason: book_status_conflict
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
      required: [reason]
...
    Book:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        title:
          type: string
        originalFilename:
          type: string
        primaryCategory:
          type: string
          enum: [course_material, research_paper, project_doc, policy_regulation, reference_book, personal_note, how_to_guide, other]
        tags:
          type: array
          items:
            type: string
        format:
          type: string
          enum: [pdf, epub, txt]
        language:
          type: string
          enum: [zh, en, other, unknown]
        status:
          type: string
          enum: [queued, processing, ready, failed]
        errorMessage:
          type: string
        sizeBytes:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        [
          id,
          ownerId,
          title,
          originalFilename,
          primaryCategory,
          tags,
          format,
          language,
          status,
          sizeBytes,
          createdAt,
          updatedAt,
        ]
...
    ListBooksResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        count:
          type: integer
      required: [items, count]
...